engine: split lib into multiple mods
